CINDER-42 _detach_volume: adding a call to _sync_ so all writes will be cached to disk before detaching the volume
This is a workaround for fixing the image dump to volumes

This commit adds python-cinderclient, which requires a new relocatable-python version; a bump commit to follow

================================================================================

Cinder OpenStack Volume Driver / CINDER-42
copying image to volume in openstack doesn't write changes to volumes

Type:              Bug                               Status:        Open                       Assignee: Guy Rozendorn
Priority:          Minor                             Resolution:    Unresolved                 Reporter: Mike Selivanov
Affects Version/s:                                   Fix Version/s: 1.1.1
Components: driver                                                                             Created: 2014-08-04 10:56
Labels:                                                                                        Updated: 2014-08-04 10:56

Issue Links:

Sub-Tasks:

Description:
The RDO instance deployed a qemu image to Glance, and when a cinder-volume image_utils uses qemu binaries to convert the image to raw format and write it to the block device, the writes are not synced to disk and the native volume driver implementation doesn't sync the disks before detaching them.

This is a bug in Cinder but we can workaround it in our driver.

Comments: